Fix bug with a tag href
Recently found this while working with different pages, *occasionally* page links will redirect back to current page instead of intended behaviour. 
This is because <a href=""> will often redirect back to the current page, and when no pageLink attribute is provided, the default will be this.
The workaround I found to this issue was to simply specify pageLink to be some non-navigating link, such as '#' or 'javascript:void(0)'.
To fix this, just check for pageLink before printing a href.